vision/requirements.txt
absl-py==0.1.13
appnope==0.1.0
astor==0.6.2
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
boto3==1.7.4
botocore==1.10.4
configparser==3.5.0
cycler==0.10.0
decorator==4.3.0
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gast==0.2.0
grpcio==1.10.1
h5py==2.8.0rc1
html5lib==0.9999999
ipykernel==4.8.2
ipython==5.6.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.5
kiwisolver==1.0.1
linecache2==1.0.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.1
numpy==1.14.2
opencv-python==3.4.0.12
pandocfilters==1.4.2
pathlib2==2.3.2
pbr==4.0.2
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.4
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
s3transfer==0.1.13
sagemaker==1.2.3
scandir==1.7
scipy==1.0.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.2.7
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traceback2==1.4.0
traitlets==4.3.2
unittest2==1.1.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.2.1
xmltodict==0.11.0
